remove polarizer planetype, check against PolarizationOpticalElement instead
